requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.4.1
    # via django
async-timeout==4.0.2
    # via redis
billiard==4.1.0
    # via celery
celery==5.3.1
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.0
    # via django-celery-beat
deprecated==1.2.13
    # via redis
django==4.0.4
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
    #   django-filter
    #   django-js-asset
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
django-celery-beat==2.5.0
    # via -r requirements.in
django-celery-results==2.5.1
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-js-asset==2.1.0
    # via django-mptt
django-mptt==0.14.0
    # via -r requirements.in
django-redis==5.3.0
    # via -r requirements.in
django-timezone-field==5.1
    # via django-celery-beat
djangorestframework==3.14.0
    # via -r requirements.in
kombu==5.3.1
    # via celery
markdown==3.4.4
    # via -r requirements.in
packaging==21.3
    # via redis
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.7
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   celery
    #   python-crontab
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via
    #   django-timezone-field
    #   djangorestframework
redis==4.3.1
    # via
    #   -r requirements.in
    #   django-redis
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
tzdata==2023.3
    # via
    #   celery
    #   django-celery-beat
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated
